Renaming "speaker" functions to "buzzer"
