+++ Gizmo/src/user_interface/src/buzzer/buzzer.ino
@@ -1,0 +1,12 @@
+#include "buzzer_class.h"
+
+buzzer_class buzzer;
+void setup() {
+  buzzer.victoryTune();
+  delay(1000);
+  buzzer.touchTune();
+}
+
+void loop() {
+  // Empty Loop
+}
