--- Gizmo/src/user_interface/src/speaker/speaker.ino
@@ -1,12 +1,0 @@
-#include "speaker_class.h"
-
-speaker_class speaker;
-void setup() {
-  speaker.victoryTune();
-  delay(1000);
-  speaker.touchTune();
-}
-
-void loop() {
-  // Empty Loop
-}
