+++ Gizmo/src/user_interface/user_interface_class.h
@@ -4,7 +4,7 @@
 
   #include "src/oled_screen/oled_screen_class.h"
   #include "src/ui_button/ui_button_class.h"
-  #include "src/speaker/speaker_class.h"
+  #include "src/buzzer/buzzer_class.h"
